Split 2FA controller between SMS and APP methods
Better to have some duplicated code in two different controllers that
are quite simple than having an unnecessary abstraction caused by
fighting the Rails way to do things.

Namespaced 2FA forms under SecondaryAuthentication module so now
the file structure and routes matches the secondary authentication
domain.

SecondaryAuthentication module allowed us to shorten the calls
between classes defined in the same module.
