+++ cosmetics-web/config/routes.rb
@@ -16,14 +16,19 @@ Rails.application.routes.draw do
   resource :password_changed, controller: "users/password_changed", only: :show, path: "password-changed"
 
   scope module: "secondary_authentication", path: "two-factor" do
-    get "auth", to: "secondary_authentication#new", as: :new_secondary_authentication
-    post "auth", to: "secondary_authentication#create", as: :secondary_authentication
-
-    get "text-not-received", to: "resend_code#new", as: :new_resend_secondary_authentication_code
-    post "text-not-received", to: "resend_code#create", as: :resend_secondary_authentication_code
-
     get "method", to: "method#new", as: :new_secondary_authentication_method
     post "method", to: "method#create", as: :secondary_authentication_method
+
+    get "sms", to: "sms#new", as: :new_secondary_authentication_sms
+    post "sms", to: "sms#create", as: :secondary_authentication_sms
+
+    get "app", to: "app#new", as: :new_secondary_authentication_app
+    post "app", to: "app#create", as: :secondary_authentication_app
+
+    scope module: "sms", path: "sms" do
+      get "not-received", to: "resend#new", as: :new_secondary_authentication_sms_resend
+      post "not-received", to: "resend#create", as: :secondary_authentication_sms_resend
+    end
   end
 
   unless Rails.env.production? && (!ENV["SIDEKIQ_USERNAME"] || !ENV["SIDEKIQ_PASSWORD"])
