+++ cosmetics-web/spec/forms/secondary_authentication/sms/resend_form_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe ResendSecondaryAuthenticationCodeForm, :with_stubbed_mailer, :with_2fa do
+RSpec.describe SecondaryAuthentication::Sms::ResendForm, :with_stubbed_mailer, :with_2fa do
   subject(:form) { described_class.new(params) }
 
   let(:user) { create(:submit_user, has_viewed_introduction: true) }
