+++ cosmetics-web/spec/forms/secondary_authentication/sms_form_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe SecondaryAuthenticationWithSmsForm, :with_stubbed_notify do
+RSpec.describe SecondaryAuthentication::SmsForm, :with_stubbed_notify do
   subject(:form) { described_class.new(otp_code: otp_code, user_id: user.id) }
 
   let(:user) { create(:submit_user, second_factor_attempts_count: attempts, direct_otp_sent_at: direct_otp_sent_at) }
