+++ cosmetics-web/spec/requests/secondary_authentication/user_requests_new_secondary_authentication_sms_code_spec.rb
@@ -1,12 +1,12 @@
 require "rails_helper"
 
-RSpec.describe "User requests new secondary authentication code", type: :request, with_stubbed_notify: true, with_2fa: true do
+RSpec.describe "User requests new secondary authentication sms code", type: :request, with_stubbed_notify: true, with_2fa: true do
   before do
     configure_requests_for_submit_domain
   end
 
   describe "viewing the form" do
-    subject(:request_code) { get new_resend_secondary_authentication_code_path }
+    subject(:request_code) { get new_secondary_authentication_sms_resend_path }
 
     context "with an user session" do
       let(:user) { create(:submit_user, :with_responsible_person, :with_sms_secondary_authentication) }
@@ -19,7 +19,7 @@ RSpec.describe "User requests new secondary authentication code", type: :request
         request_code
 
         expect(response).to have_http_status(:ok)
-        expect(response).to render_template("secondary_authentication/resend_code/new")
+        expect(response).to render_template("secondary_authentication/sms/resend/new")
       end
     end
 
@@ -36,7 +36,7 @@ RSpec.describe "User requests new secondary authentication code", type: :request
   describe "submitting the request" do
     context "without an user session" do
       it "shows an access denied error", :aggregate_failures do
-        post resend_secondary_authentication_code_path
+        post secondary_authentication_sms_resend_path
 
         expect(response).to have_http_status(:forbidden)
         expect(response).to render_template("errors/forbidden")
@@ -44,7 +44,7 @@ RSpec.describe "User requests new secondary authentication code", type: :request
     end
 
     context "with an user session" do
-      subject(:request_code) { post new_resend_secondary_authentication_code_path }
+      subject(:request_code) { post new_secondary_authentication_sms_resend_path }
 
       let(:user) { create(:submit_user, :with_responsible_person, :with_sms_secondary_authentication) }
 
@@ -70,13 +70,13 @@ RSpec.describe "User requests new secondary authentication code", type: :request
       it "redirects the user to the secondary authentication page" do
         request_code
 
-        expect(response).to redirect_to(new_secondary_authentication_path)
+        expect(response).to redirect_to(new_secondary_authentication_sms_path)
       end
     end
 
     context "with an user session corresponding to an user who haven't verified their mobile number" do
       subject(:request_code) do
-        post new_resend_secondary_authentication_code_path, params: { submit_user: { mobile_number: mobile_number } }
+        post new_secondary_authentication_sms_resend_path, params: { submit_user: { mobile_number: mobile_number } }
       end
 
       let(:user) do
@@ -93,7 +93,7 @@ RSpec.describe "User requests new secondary authentication code", type: :request
         it "shows the submission form" do
           request_code
 
-          expect(response).to render_template("secondary_authentication/resend_code/new")
+          expect(response).to render_template("secondary_authentication/sms/resend/new")
         end
 
         it "does not change the user secondary authentication code" do
@@ -116,7 +116,7 @@ RSpec.describe "User requests new secondary authentication code", type: :request
         it "shows the submission form" do
           request_code
 
-          expect(response).to render_template("secondary_authentication/resend_code/new")
+          expect(response).to render_template("secondary_authentication/sms/resend/new")
         end
 
         it "does not change the user secondary authentication code" do
@@ -161,7 +161,7 @@ RSpec.describe "User requests new secondary authentication code", type: :request
         it "redirects the user to the secondary authentication page" do
           request_code
 
-          expect(response).to redirect_to(new_secondary_authentication_path)
+          expect(response).to redirect_to(new_secondary_authentication_sms_path)
         end
       end
     end
